+++ sdk/@launchdarkly/session-replay/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @launchdarkly/session-replay
 
+## 0.3.10
+
+### Patch Changes
+
+- Updated dependencies [1beae4d]
+    - highlight.run@9.19.1
+
 ## 0.3.9
 
 ### Patch Changes