+++ sdk/highlight-nest/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @highlight-run/nest
 
+## 3.6.33
+
+### Patch Changes
+
+- @highlight-run/node@3.12.21
+
 ## 3.6.32
 
 ### Patch Changes